Move block tests into components
